fix:anthropic sse
